#39631 Move AuditLogger into BusinessLogic project
